+++ posthog/src/main/java/com/posthog/internal/surveys/GsonSurveyScheduleAdapter.kt
@@ -1,0 +1,35 @@
+package com.posthog.internal.surveys
+
+import com.google.gson.JsonDeserializationContext
+import com.google.gson.JsonDeserializer
+import com.google.gson.JsonElement
+import com.google.gson.JsonSerializationContext
+import com.google.gson.JsonSerializer
+import com.posthog.PostHogConfig
+import com.posthog.surveys.SurveySchedule
+import java.lang.reflect.Type
+
+internal class GsonSurveyScheduleAdapter(private val config: PostHogConfig) :
+    JsonSerializer<SurveySchedule>,
+    JsonDeserializer<SurveySchedule> {
+    override fun serialize(
+        src: SurveySchedule,
+        typeOfSrc: Type,
+        context: JsonSerializationContext,
+    ): JsonElement {
+        return context.serialize(src.value)
+    }
+
+    override fun deserialize(
+        json: JsonElement,
+        typeOfT: Type,
+        context: JsonDeserializationContext,
+    ): SurveySchedule? {
+        return try {
+            return SurveySchedule.fromValue(json.asString)
+        } catch (e: Throwable) {
+            config.logger.log("${json.asString} isn't a known type: $e.")
+            null
+        }
+    }
+}